+++ lib/ReactViews/Search/SidebarSearch.jsx
@@ -1,12 +1,9 @@
 import ObserveModelMixin from '../ObserveModelMixin';
 import React from 'react';
-import SearchHeader from './SearchHeader.jsx';
 import SearchResult from './SearchResult.jsx';
 import BadgeBar from '../BadgeBar.jsx';
-import classNames from 'classnames';
-import Icon from "../Icon.jsx";
-
 import Styles from './sidebar-search.scss';
+import LocationSearchResults from './LocationSearchResults.jsx';
 
 import {addMarker} from './SearchMarkerUtils';
 
@@ -28,9 +25,13 @@ const SidebarSearch = React.createClass({
         this.props.viewState.searchState.showLocationSearchResults = false;
     },
 
+    onLocationClick(result) {
+        addMarker(this.props.terria, this.props.viewState, result);
+        result.clickAction();
+    },
+
     render() {
         const searchResultCount = this.props.viewState.searchState.locationSearchProviders.reduce((count, result) => count + result.searchResults.length, 0);
-
         return (
             <div className={Styles.search}>
                 <div className={Styles.results}>
@@ -45,6 +46,7 @@ const SidebarSearch = React.createClass({
                                                    terria={this.props.terria}
                                                    viewState={this.props.viewState}
                                                    search={search}
+                                                   onLocationClick={this.onLocationClick}
                                                    isWaitingForSearchToStart={this.props.isWaitingForSearchToStart}
 
                             />
@@ -64,51 +66,6 @@ const SidebarSearch = React.createClass({
                 </div>
             </div>
         );
-    }
-});
-
-const LocationSearchResults = React.createClass({
-    propTypes: {
-        viewState: React.PropTypes.object.isRequired,
-        isWaitingForSearchToStart: React.PropTypes.bool,
-        terria: React.PropTypes.object.isRequired,
-        search: React.PropTypes.object.isRequired
-    },
-
-    getInitialState() {
-        return {
-            isOpen: true
-        };
-    },
-
-    toggleGroup() {
-        this.setState({
-            isOpen: !this.state.isOpen
-        });
-    },
-
-    onLocationClick(result) {
-        addMarker(this.props.terria, this.props.viewState, result);
-        result.clickAction();
-    },
-
-    render() {
-        const search = this.props.search;
-        return (<div key={search.name} className={classNames(Styles.providerResult, {[Styles.isOpen]: this.state.isOpen})}>
-                    <button onClick={this.toggleGroup} className={Styles.heading}>
-                        <span>{search.name}</span>
-                        <Icon glyph={this.state.isOpen ? Icon.GLYPHS.opened : Icon.GLYPHS.closed}/>
-                    </button>
-                    <SearchHeader searchProvider={search}
-                                  isWaitingForSearchToStart={this.props.isWaitingForSearchToStart}/>
-                    <ul className={Styles.items}>
-                        {search.searchResults.map((result, i) => (
-                            <SearchResult key={i}
-                                          clickAction={this.onLocationClick.bind(this, result)}
-                                          name={result.name}/>
-                        ))}
-                    </ul>
-                </div>);
     }
 });
